Increase fcnt_exp to 32 bit to allow clk freq checking. Got:
==== Marble Init      ====  : PASS.
  Fclk  DSP_CLK: 119.001 MHz
==== ZEST DSP CLK Freq====  : PASS.
==== ZEST ADC(AD9653) ====  : PASS.
==== ZEST DAC(AD9781) ====  : PASS.
==== ZEST AD7794      ====  : PASS.
==== ZEST AMC7823     ====  : PASS.
  Fclk ADC0_DIV: 118.999 MHz
  Clk DIV freq  0 Check: PASS.
    Phase ADC0_DIV clk: 0x30 0.375 UI
  Phase ADC0_DIV clk aligned. retry = 0.
  Clk DIV phase 0 Check: PASS.
  Fclk ADC1_DIV: 118.999 MHz
  Clk DIV freq  1 Check: PASS.
    Phase ADC1_DIV clk: 0x37 0.429 UI
  Phase ADC1_DIV clk aligned. retry = 0.
  Clk DIV phase 1 Check: PASS.
  Fclk  DAC_DCO: 237.998 MHz
  Clk DIV freq  2 Check: PASS.

+++ soc/marble_zest/common/ui.c
@@ -15,7 +15,7 @@
 extern t_zest_init zest_init_data;
 
 void handle_ui( void ) {
-    uint16_t *fcnt_exp = zest_init_data.fcnt_exp;
+    uint32_t *fcnt_exp = zest_init_data.fcnt_exp;
     uint8_t *phs_center = zest_init_data.phs_center;
     uint16_t tempC = UART_GETC( BASE_UART0 );
     int16_t dval;
